<REFACTOR> LottoNumbers class 없애고 Lotto 로 위임하기
