+++ src/main/java/domain/LottoNumbers.java
@@ -1,38 +1,38 @@
-package domain;
-
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-
-public class LottoNumbers {
-    public static final int SIZE = 6;
-    private final List<LottoNumber> numbers;
-
-    public List<LottoNumber> getNumbers() {
-        return numbers;
-    }
-
-    public LottoNumbers(List<LottoNumber> numbers) {
-        validateNumber(numbers);
-        this.numbers = numbers;
-    }
-
-    private void validateNumber(List<LottoNumber> numbers) {
-        validateSize(numbers);
-        validateDuplicate(numbers);
-    }
-
-    private void validateSize(List<LottoNumber> numbers) {
-        if (numbers.size() != SIZE) {
-            throw new IllegalArgumentException("로또 번호의 갯수는 6이여야 합니다.");
-        }
-    }
-
-    private void validateDuplicate(List<LottoNumber> numbers) {
-        Set<LottoNumber> setNumbers = new HashSet<>(numbers);
-
-        if (setNumbers.size() != numbers.size()) {
-            throw new IllegalArgumentException("로또 번호는 중복이 없어야 합니다");
-        }
-    }
-}
+//package domain;
+//
+//import java.util.HashSet;
+//import java.util.List;
+//import java.util.Set;
+//
+//public class LottoNumbers {
+////    public static final int SIZE = 6;
+//    private final List<LottoNumber> numbers;
+//
+//    public List<LottoNumber> getNumbers() {
+//        return numbers;
+//    }
+//
+//    public LottoNumbers(List<LottoNumber> numbers) {
+//        validateNumber(numbers);
+//        this.numbers = numbers;
+//    }
+//
+//    private void validateNumber(List<LottoNumber> numbers) {
+//        validateSize(numbers);
+//        validateDuplicate(numbers);
+//    }
+//
+//    private void validateSize(List<LottoNumber> numbers) {
+//        if (numbers.size() != SIZE) {
+//            throw new IllegalArgumentException("로또 번호의 갯수는 6이여야 합니다.");
+//        }
+//    }
+//
+//    private void validateDuplicate(List<LottoNumber> numbers) {
+//        Set<LottoNumber> setNumbers = new HashSet<>(numbers);
+//
+//        if (setNumbers.size() != numbers.size()) {
+//            throw new IllegalArgumentException("로또 번호는 중복이 없어야 합니다");
+//        }
+//    }
+//}
